Integrated Ghoulslash's GFX ID expansion bumping OW sprite limit to u16

+++ src/cable_car.c
@@ -800,12 +800,12 @@ static void CreateCableCarSprites(void)
     u8 spriteId;
     u8 i;
 
-    u8 playerGraphicsIds[2] = {
+    u16 playerGraphicsIds[2] = {
         [STYLE_EMR_BRENDAN] = OBJ_EVENT_GFX_RIVAL_BRENDAN_NORMAL,
         [STYLE_EMR_MAY] = OBJ_EVENT_GFX_RIVAL_MAY_NORMAL
     };
     u16 rval = Random();
-    u8 hikerGraphicsIds[4] = {
+    u16 hikerGraphicsIds[4] = {
         OBJ_EVENT_GFX_HIKER,
         OBJ_EVENT_GFX_CAMPER,
         OBJ_EVENT_GFX_PICNICKER,
